Use CancelIoEx in case of async exceptions
If a thread which is blocked on reading or writing is cancelled with an
'AsyncException' use 'CancelIoEx' to cancel the IO action.  When

'CancelIoEx' is used the 'IOManager' thread is notified and it will
deallocate 'ioDataPtr' which was allocated on process heap.  In such
a case 'GetQueuedCompletionStatus' returns 'False' and sets
'ERROR_OPERATION_ABORTED' error.